Fix infinite loop in DeltaCompressedReaderMapper.
The trait std::io::Read, (in the current Rust version at least - 1.6), defines
`read` to return Ok(0) when "end of file" (EOF) has been reached. This causes
the main loop in DeltaCompressedReaderMapper::map_edges to never terminate.

This commit simply adds a check on the number of bytes read. If it is zero, we
assume that EOF has been reached and we break out of the loop.